+++ src/Network/Network/Cortex/VirtualHub/NewAzureRmVirtualHubCommand.cs
@@ -125,6 +125,11 @@ namespace Microsoft.Azure.Commands.Network
         public uint VirtualRouterAsn { get; set; }
 
         [Parameter(
+        Mandatory = false,
+        HelpMessage = "Autoscale configuration for the hub router")]
+        public PSVirtualRouterAutoScaleConfiguration VirtualRouterAutoScaleConfiguration { get; set; }
+
+        [Parameter(
             Mandatory = false,
             HelpMessage = "Run cmdlet in the background")]
         public SwitchParameter AsJob { get; set; }
@@ -175,7 +180,8 @@ namespace Microsoft.Azure.Commands.Network
                         VirtualWan = new PSResourceId() { Id = resolvedVirtualWan.Id },
                         AddressPrefix = this.AddressPrefix,
                         Location = this.Location,
-                        VirtualRouterAsn = this.VirtualRouterAsn
+                        VirtualRouterAsn = this.VirtualRouterAsn,
+                        VirtualRouterAutoScaleConfiguration = this.VirtualRouterAutoScaleConfiguration
                     };
 
                     virtualHub.RouteTable = this.RouteTable;